Add payment settlement_risk field to relevant PaymentDetail classes
