+++ src/main/java/com/truelayer/java/payments/entities/paymentdetail/ExecutedPaymentDetail.java
@@ -19,8 +19,15 @@ public class ExecutedPaymentDetail extends PaymentDetail {
 
     AuthorizationFlowWithConfiguration authorizationFlow;
 
+    SettlementRisk settlementRisk;
+
     @JsonGetter
     public Optional<AuthorizationFlowWithConfiguration> getAuthorizationFlow() {
         return Optional.ofNullable(authorizationFlow);
+    }
+
+    @JsonGetter
+    public Optional<SettlementRisk> getSettlementRisk() {
+        return Optional.ofNullable(settlementRisk);
     }
 }
